SearchAddress: move to Geotag subdirectory
